Move KafkaProperties out of autoconfig directory
Move KafkaProperties out of autoconfig directory

#754
