+++ spring-boot-autoconfigure/src/main/java/org/axonframework/boot/KafkaProperties.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.axonframework.boot.autoconfig;
+package org.axonframework.boot;
 
 import org.apache.kafka.clients.CommonClientConfigs;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
